backend/requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
sqlalchemy==2.0.*
supabase==2.*
python-dotenv==1.0.*
pydantic==2.*
pydantic-settings==2.*
alembic==1.*
stripe==11.*
twilio==9.*
jinja2==3.*
httpx==0.*
weasyprint==62.*
python-multipart==0.0.*
PyJWT==2.*
